Use weak values for comprehensions (#184)
I found that it's fairly common to mutate containers created using comprehensions later, so allow it. This also moves the "weak" property to the container instead of the generic parameter, and adds an implementation for dict.setdefault.

This uncovered a bug in the handling of dictionary comprehensions that I will fix separately.