Add --disable-all CLI flag and allow enabling/disabling specific checks when --*-all passed (#215)
Fixes #211 

Allows checks to be disabled with `-d` when `--enable-all` is passed, to support a strict default-enabled/deny-list approach to checks. Furthermore, adds a new `--disable-all` flag to the CLI, which disables all checks except those enabled with `-e`, to support a default-disabled/allow-list approach.

There didn't appear to be an obvious one-letter equivalent (`-A` and `-a` were already taken, as was `-d`, and `-D` was potentially confusingly not parallel with `-a`. However, if you want me to add one, I certainly can.