add concept of AnySource (#251)
I'd been thinking of doing this for longer, but now I have a motivating use case. A CustomCheck that excludes Any may still want to allow one specific unavoidable kind of Any: the type of the element of an empty list. To support that, we replace the singleton `UNRESOLVED_VALUE` with instances of a new class `AnyValue` that provides a `source`.

This PR merely creates the enum and replaces `is UNRESOLVED_VALUE` checks within pyanalyze. After I migrate some internal code, I will replace usages of UNRESOLVED_VALUE within pyanalyze.