Remove WeakExtension (#517)
With the new support for variable-length elements in SequenceValue,
we can use a SequenceValue with a single variable-length element
instead of a "weak" list. The behavior is roughly equivalent, but allows
more precise inference in some cases.

We add an impl for dict.pop to compensate for some of this more
precise inference.